# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Travel Assistant API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/plans/parse": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Parse Plan Text",
        "description": "Parse a freeform text input and return a TripPlanCreateRequest.\n\nSimple heuristics to extract origin, destination, dates, days and budget.\nIf no explicit start date is found, default to today.",
        "operationId": "parse_plan_text_api_plans_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlanCreateRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Plan Task",
        "operationId": "create_plan_task_api_plans_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPlanCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanTaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_api_plans_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TripPlanResponse"
                  },
                  "title": "Response List Plans Api Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_api_plans__plan_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/summary": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan Summary",
        "operationId": "get_plan_summary_api_plans__plan_id__summary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/versions": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Plan Versions",
        "operationId": "list_plan_versions_api_plans__plan_id__versions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TripPlanVersionResponse"
                  },
                  "title": "Response List Plan Versions Api Plans  Plan Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/versions/{version_id}/regenerate": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Regenerate Plan",
        "operationId": "regenerate_plan_api_plans__plan_id__versions__version_id__regenerate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPlanCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanTaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/versions/{version_id}": {
      "put": {
        "tags": [
          "plans"
        ],
        "summary": "Edit Plan Version",
        "operationId": "edit_plan_version_api_plans__plan_id__versions__version_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPlanEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlanVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/versions/{version_id}/restore": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Restore Plan Version",
        "operationId": "restore_plan_version_api_plans__plan_id__versions__version_id__restore_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlanVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/versions/{version_id}/export": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Export Plan Version Pdf",
        "operationId": "export_plan_version_pdf_api_plans__plan_id__versions__version_id__export_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/warnings": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan Warnings",
        "operationId": "get_plan_warnings_api_plans__plan_id__warnings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherWarningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/versions/{version_id}/warnings": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Version Warnings",
        "operationId": "get_version_warnings_api_plans__plan_id__versions__version_id__warnings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherWarningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParseRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ParseRequest"
      },
      "PlanSummaryResponse": {
        "properties": {
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "budget_range": {
            "type": "string",
            "title": "Budget Range"
          },
          "current_version_id": {
            "type": "integer",
            "title": "Current Version Id"
          },
          "current_version_no": {
            "type": "integer",
            "title": "Current Version No"
          },
          "estimated_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Total"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "pace": {
            "type": "string",
            "title": "Pace"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "title",
          "city",
          "start_date",
          "end_date",
          "budget_range",
          "current_version_id",
          "current_version_no",
          "estimated_total",
          "risk_level",
          "pace",
          "warning_count",
          "updated_at"
        ],
        "title": "PlanSummaryResponse"
      },
      "PlanTaskCreateResponse": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "PlanTaskCreateResponse"
      },
...
      "TripPlanCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "City"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "budget_range": {
            "type": "string",
            "title": "Budget Range"
          },
          "transport_preference": {
            "type": "string",
            "title": "Transport Preference",
            "default": "public_transit"
          },
          "accommodation_preference": {
            "type": "string",
            "title": "Accommodation Preference",
            "default": "comfort"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "title",
          "city",
          "start_date",
          "end_date",
          "budget_range"
        ],
        "title": "TripPlanCreateRequest"
      },
      "TripPlanEditRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary",
            "default": "edited manually"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "TripPlanEditRequest"
      },
      "TripPlanResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "owner_user_id": {
            "type": "integer",
            "title": "Owner User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "budget_range": {
            "type": "string",
            "title": "Budget Range"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "current_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripPlanVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_user_id",
          "title",
          "city",
          "start_date",
          "end_date",
          "budget_range",
          "current_version_id",
          "created_at",
          "updated_at"
        ],
        "title": "TripPlanResponse"
      },
      "TripPlanVersionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "parent_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version Id"
          },
          "owner_user_id": {
            "type": "integer",
            "title": "Owner User Id"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary"
          },
          "content_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "parent_version_id",
          "owner_user_id",
          "version_no",
          "source_type",
          "change_summary",
          "content_json",
          "created_at"
        ],
        "title": "TripPlanVersionResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}